+++ packages/react-kit/src/components/modal/components/Redeem/ExchangeView/cancellation/CancelExchange.tsx
@@ -250,7 +250,7 @@ export function CancelExchange({
               toast((t) => (
                 <SuccessTransactionToast
                   t={t}
-                  action={`Cancelled exchange: ${offer.metadata.name}`}
+                  action={`Cancelled exchange: ${offer.metadata?.name}`}
                 />
               ));
             }}