+++ packages/react-kit/src/hooks/contracts/erc721/useErc721OwnerOf.ts
@@ -1,14 +1,13 @@
+import { CoreSDK } from "@bosonprotocol/core-sdk";
 import { useQuery } from "react-query";
-import { useCoreSDKWithContext } from "../../core-sdk/useCoreSdkWithContext";
 
 export const useErc721OwnerOf = (
   props: {
     contractAddress: string | undefined;
     tokenIds: (string | null | undefined)[] | undefined;
   },
-  { enabled }: { enabled: boolean | undefined }
+  { enabled, coreSDK }: { enabled: boolean | undefined; coreSDK: CoreSDK }
 ) => {
-  const coreSDK = useCoreSDKWithContext();
   return useQuery(
     ["erc721-owner-of", coreSDK.uuid, props],
     async () => {