+++ packages/react-kit/src/lib/bundle/const.ts
@@ -1,0 +1,18 @@
+export const digitalTypeMappingDisplay = {
+  "digital-nft": "Digital NFT",
+  "digital-file": "Digital File",
+  experiential: "Experiential"
+} as const;
+
+export const digitalTypeMapping = {
+  "digital-nft": "digital-nft",
+  "digital-file": "digital-file",
+  experiential: "experiential"
+} as const;
+
+export const digitalNftTypeMapping = {
+  wearable: "Wearable",
+  image: "Image",
+  event: "Event",
+  other: "Other"
+};