+++ packages/react-kit/src/types/variants.ts
@@ -4,8 +4,7 @@ import { Offer } from "./offer";
 
 export type VariantV1 = {
   offer: Offer;
-  variations: (Omit<subgraph.ProductV1Variation, "metadata" | "type"> & {
-    metadata?: subgraph.ProductV1MetadataEntity;
+  variations: (Omit<subgraph.ProductV1Variation, "type"> & {
     type: "Color" | "Size";
   })[];
 };