+++ e2e/meta-tx-gateway/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.3](https://github.com/bosonprotocol/core-components/compare/meta-tx-gateway@1.7.2...meta-tx-gateway@1.7.3) (2024-03-08)
+
+**Note:** Version bump only for package meta-tx-gateway
+
+
+
+
+
 ## [1.7.2](https://github.com/bosonprotocol/core-components/compare/meta-tx-gateway@1.7.1...meta-tx-gateway@1.7.2) (2024-02-23)
 
 **Note:** Version bump only for package meta-tx-gateway